feat(home): apply explore-style filter + toolbar layout
- Extract FilterPanel / Toolbar / CourseCard into shared _components/
  so home and explore render the same browse surface.
- Include author info in findAllPublishedWithLessons so the home card
  can show author name + avatar like explore.
- Home uses a "おすすめ" sort default (disabled preview) to mark the
  curated intent vs. explore's community-sort options.
- Brand link in TopBar now targets /explore so the default landing
  page emphasises community content.

Closes part of #53.
